Get information tokens and improve the types
