+++ src/api/tenderly/index.ts
@@ -1,1 +1,3 @@
-export * from './TenderlyApi'
+export * from './tenderlyApi'
+
+export { PublicTrade as Trade, Transfer, Account } from './types'
